+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## 0.0.2
+- ENTER key in file name text box selects template
+- Template selector remembers last selected template
+- getImputNameVariable was returning empy string instead of $itemname$ variable
+
 ## 0.0.1
 - Initial release